requirements.txt
ultralytics>=8.3.0
albumentations>=1.4.21
fastapi>=0.116.0
uvicorn[standard]>=0.35.0
python-multipart>=0.0.20
numpy>=2.2.0
opencv-python>=4.11.0
Pillow>=11.0.0
PyYAML>=6.0.2
pycocotools>=2.0.8
